Fix auditclient location
