+++ modules/basis/basis-auditlog/src/main/java/com/sap/xsk/auditlog/client/messages/ConfigurationChange.java
